ENH: Restore filler
Swap double filler for old version
Restore fillers to distributions
Replace pointer size
Switch to uintptr_t for 32 bit platforms

+++ _randomgen/randomgen/pcg32.pyx
@@ -345,18 +345,18 @@ cdef class PCG32:
 
         import ctypes
 
-        self._ctypes = interface(<Py_ssize_t>self.rng_state,
-                         ctypes.c_void_p(<Py_ssize_t>self.rng_state),
-                         ctypes.cast(<Py_ssize_t>&pcg32_uint64,
+        self._ctypes = interface(<uintptr_t>self.rng_state,
+                         ctypes.c_void_p(<uintptr_t>self.rng_state),
+                         ctypes.cast(<uintptr_t>&pcg32_uint64,
                                      ctypes.CFUNCTYPE(ctypes.c_uint64,
                                      ctypes.c_void_p)),
-                         ctypes.cast(<Py_ssize_t>&pcg32_uint32,
+                         ctypes.cast(<uintptr_t>&pcg32_uint32,
                                      ctypes.CFUNCTYPE(ctypes.c_uint32,
                                      ctypes.c_void_p)),
-                         ctypes.cast(<Py_ssize_t>&pcg32_double,
+                         ctypes.cast(<uintptr_t>&pcg32_double,
                                      ctypes.CFUNCTYPE(ctypes.c_double,
                                      ctypes.c_void_p)),
-                         ctypes.c_void_p(<Py_ssize_t>self._brng))
+                         ctypes.c_void_p(<uintptr_t>self._brng))
         return self.ctypes
 
     @property
@@ -384,12 +384,12 @@ cdef class PCG32:
             raise ImportError('cffi is cannot be imported.')
 
         ffi = cffi.FFI()
-        self._cffi = interface(<Py_ssize_t>self.rng_state,
-                         ffi.cast('void *',<Py_ssize_t>self.rng_state),
-                         ffi.cast('uint64_t (*)(void *)',<uint64_t>self._brng.next_uint64),
-                         ffi.cast('uint32_t (*)(void *)',<uint64_t>self._brng.next_uint32),
-                         ffi.cast('double (*)(void *)',<uint64_t>self._brng.next_double),
-                         ffi.cast('void *',<Py_ssize_t>self._brng))
+        self._cffi = interface(<uintptr_t>self.rng_state,
+                         ffi.cast('void *',<uintptr_t>self.rng_state),
+                         ffi.cast('uint64_t (*)(void *)',<uintptr_t>self._brng.next_uint64),
+                         ffi.cast('uint32_t (*)(void *)',<uintptr_t>self._brng.next_uint32),
+                         ffi.cast('double (*)(void *)',<uintptr_t>self._brng.next_double),
+                         ffi.cast('void *',<uintptr_t>self._brng))
         return self.cffi
 
     @property
